refactor(server): extract device config into core microcrate
