+++ crates/bitnet-device-config-core/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "bitnet-device-config-core"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "SRP core device configuration parsing and resolution"
+keywords.workspace = true
+categories.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+bitnet-common = { path = "../bitnet-common", version = "0.2.1-dev" }
+bitnet-kernels = { path = "../bitnet-kernels", version = "0.2.1-dev", optional = true }
+serde.workspace = true
+
+[features]
+default = []
+cuda = ["dep:bitnet-kernels", "bitnet-kernels/cuda"]
+gpu = ["cuda", "bitnet-kernels/gpu"]
