+++ crates/bitnet-server/Cargo.toml
@@ -22,6 +22,7 @@ async-stream = "0.3.6"
 async-trait = "0.1.89"
 axum.workspace = true
 bitnet-common = { path = "../bitnet-common", version = "0.2.1-dev" }
+bitnet-device-config-core = { path = "../bitnet-device-config-core", version = "0.2.1-dev" }
 bitnet-eval-core = { path = "../bitnet-eval-core", version = "0.2.1-dev" }
 bitnet-inference = { path = "../bitnet-inference", version = "0.2.1-dev" }
 bitnet-kernels = { path = "../bitnet-kernels", version = "0.2.1-dev" }
@@ -93,6 +94,7 @@ cpu = [
 ]
 cuda = [
   "dep:cudarc",
+  "bitnet-device-config-core/cuda",
   "bitnet-inference/cuda",
   "bitnet-kernels/cuda",
   "runtime-profile-core",
@@ -103,6 +105,7 @@ default = ["prometheus", "server"]
 degraded-ok = []                                                                                                                                                                  # Map Degraded → 200 OK (default is fail-fast)
 gpu = [
   "cuda",
+  "bitnet-device-config-core/gpu",
   "bitnet-runtime-bootstrap/runtime-profile-gpu",
   "runtime-profile-core",
 ]
